[ALPS04867758] USB: Fix SSUSB enumeration issue with multiple interfaces
[Detail]
1. Allocate single buffer for all EPs at SuperSpeed in saving mode.
2. Fixed EP0 TX count variable type from u8 to u16.

MTK-Commit-Id: 88544940573474b0fb3e626cf7cc14ab84869e2a

CR-Id: ALPS04867758
Feature: USB 3.0 Device

+++ drivers/usb/mtu3/mtu3_gadget_ep0.c
@@ -560,7 +560,7 @@ static void ep0_tx_state(struct mtu3 *mtu)
 	struct usb_request *req;
 	u32 csr;
 	u8 *src;
-	u8 count;
+	u16 count;
 	u32 maxp;
 
 	dev_dbg(mtu->dev, "%s\n", __func__);
